deployment/requirements.txt
# Core API Framework
fastapi
uvicorn[standard]
pydantic

# Machine Learning
scikit-learn
joblib
numpy
pandas
xgboost

# Database
psycopg2-binary

# Data Visualization & Analysis
scipy
seaborn
plotly

# Data enrichment (for fuzzy matching if needed)
rapidfuzz

# Monitoring and logging
python-multipart
python-json-logger

# CORS and security
python-jose[cryptography]

# Testing
pytest
httpx